requirements/pypi.txt
# These dependencies are installed using:
#
#   pip install --upgrade
#
# This list is the default package list. All pure python packages
# for the production environment at runtime should be listed here.
# Binary (non-pure) packages may also be listed here, but you
# should see if there is a conda package that suits your needs.

click==8.0.1
pylev==1.3.0
argon2-cffi==20.1.0
PyQt5==5.15.4

mypy-extensions
typing-extensions
